implement a better (less buggy) and simpler Format function
Previous function has all sorts of issues with undo points, losing position, moving the cursor around, moving the window around, etc.
Implemented as a new function instead of modifying the existing one to prevent future merge conflicts as upstream has no desire to merge the changes, and I will be maintaining them on my own branch.
The new function is vastly simpler, as the old function deliberately is trying to introduce a bug where it loses the save point for the format operation.